surveillance/requirements.txt
aiosqlite==0.20.0
alembic==1.14.1
annotated-types==0.7.0
anyio==4.8.0
asyncpg==0.30.0
autopep8==2.3.2
certifi==2024.12.14
click==8.1.8
-e git+ssh://git@github.com/rolandm2017/deskSense.git@d673ea7eaefc8ef3fbb5c9cc34bc037f6ec75c21#egg=deskSense&subdirectory=surveillance
evdev==1.7.1
fastapi==0.115.6
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
Mako==1.3.8
MarkupSafe==3.0.2
packaging==24.2
pluggy==1.5.0
psutil==6.1.1
psycopg2-binary==2.9.10
pycodestyle==2.12.1
pydantic==2.10.5
pydantic_core==2.27.2
PyGetWindow==0.0.9
pynput==1.7.7
PyRect==0.2.0
pytest==8.3.4
pytest-asyncio==0.25.2
python-dotenv==1.0.1
python-xlib==0.33
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.37
starlette==0.41.3
typing_extensions==4.12.2
uvicorn==0.34.0
